@@ -1,23 +1,19 @@
-﻿using System;
-
-namespace DapperNpa.SourceGenerator;
+﻿namespace DapperNpa.SourceGenerator;
 
 internal sealed partial class RepositorySourceGenerator
 {
-    private static readonly string RepositoryImplementationClass = $@"// <auto-generated/>
-// {DateTime.Now}
+    private static string RepositoryImplementationTemplate = $$"""
+// <auto-generated/>
+// {{DateTime.Now}}
 #nullable enable
 
-namespace DapperNpa.Repository
-{{
-    internal sealed class UserRepositoryImpl : global::DapperNpa.Aspnet.Example.Repository.IUserRepository
-    {{
-        public global::DapperNpa.Aspnet.Example.Model.User GetById(global::System.Guid id)
-        {{
-            return new global::DapperNpa.Aspnet.Example.Model.User();
-        }}
-    }}
-}}
-";
+namespace __NAMESPACE__
+{
+    internal sealed class __CLASS__ : global::__INTERFACE___
+    {
+        __IMPLEMENTATION_METHODS__
+    }
+}
+""";
     
 }
