@@ -2,5 +2,5 @@
 
 internal sealed partial class RepositorySourceGenerator
 {
-    private const string RepositorySourceGeneratorName = "DapperNpa.Repository.g.cs";
+    private const string RepositoryAttributeName = "Repository";
 }
